pyproject.toml
[tool.poetry]
name = "todoist-bot"
version = "0.1.0"
description = ""
authors = ["RomanSchaposhnikov <roman.schaposhnikov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
todoist-python = "^8.1.3"
python-dotenv = "^0.19.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
notebook = "^6.4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
